\name{SOES1015-package}
\alias{SOES1015-package}
\alias{SOES1015}
\keyword{package}

\docType{package}
\title{Supporting material for SOES1015}
\description{
Datasets and dependencies for SOES1015 Statistical Computing in Marine Sciences, a 1st year module in Ocean and Earth Sciences at the University of Southampton. The datasets are available online, either through R or elsewhere as published material that accompanies the manuscripts. I will add help files describing the metadata in due course (see the relevant practical session on Blackboard), but don't hesitate to contact me if you require the accompanying metadata before that happens.
}
\details{
\tabular{ll}{
Package: \tab SOES1015\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2020-01-21\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Thomas Ezard

Maintainer: Thomas Ezard \email{t.ezard@soton.ac.uk}
}
